Add autocompletion
Cobra has built in support for autocompletion. The only needed thing is
a boilerplate to add the `completion` command which generates scripts
for popular shells.

The built in cobra autocompletion uses a hidden command called
`__completion` to do the actual look-ups. To make it work we need to
bypass `PersistentPreRunE` in the root-command when `__completion` is
called.