KM478 ✅ Adding support for using named AWS profile for authentication
This enables the user to specify the '-a aws-profile' command line argument
to log in using a named AWS profile. The AWS profile name is specified in the
AWS_PROFILE environment variable.

This allows authentication with IAM access keys specified in the global
AWS credentials file without the need for setting secrets in environment
variables, or with profiles used for AWS SSO or SAML-based login.